# hack/contrib/docker/api/html/swagger.json
{
  "swagger": "2.0",
  "paths": {
...
    "/v2/tenants/query/{tenant_name}": {
      "get": {
        "description": "get tenant resources",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "v2"
        ],
        "summary": "租户带资源列表",
        "operationId": "tenants",
        "parameters": [
          {
            "type": "string",
            "format": "string",
            "description": "123",
            "name": "tenant_name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "default": {
            "description": "统一返回格式",
            "schema": {
              "$ref": "#/responses/commandResponse"
            }
          }
        }
      }
    },
...
    "/v2/tenants/{tenant_name}/res": {
      "get": {
        "description": "get tenant resources",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "v2"
        ],
        "summary": "租户带资源单个",
        "operationId": "tenants",
        "parameters": [
          {
            "type": "string",
            "format": "string",
            "description": "123",
            "name": "tenant_name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "default": {
            "description": "统一返回格式",
            "schema": {
              "$ref": "#/responses/commandResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Values": {
      "description": "It is typically used for query parameters and form values.\nUnlike in the http.Header map, the keys in a Values map\nare case-sensitive.",
      "type": "object",
      "title": "Values maps a string key to a list of values.",
      "additionalProperties": {
        "type": "array",
        "items": {
          "type": "string"
        }
      },
      "x-go-package": "net/url"
    },
...
  },
  "responses": {
    "commandResponse": {
      "description": "CommandResponse api统一返回结构",
      "schema": {
        "type": "object",
        "properties": {
          "bean": {
            "description": "单资源实体",
            "type": "object",
            "x-go-name": "Bean"
          },
          "list": {
            "description": "资源列表",
            "type": "object",
            "x-go-name": "List"
          },
          "msg": {
            "description": "API错误信息",
            "type": "string",
            "x-go-name": "Msg"
          },
          "number": {
            "description": "数据集总数",
            "type": "integer",
            "format": "int64",
            "x-go-name": "ListAllNumber"
          },
          "page": {
            "description": "当前页码数",
            "type": "integer",
            "format": "int64",
            "x-go-name": "Page"
          },
          "validation_error": {
            "$ref": "#/definitions/Values"
          }
        }
      }
    }
  }
}